Adjust rendering reset to support fixed cell grids
Responsive by default led to problems controlling the dimensions of grid
cells. This change provides an alternative option where the overall
canvas width is automatically determined by multiplying the cell width
by the number of cells in each axis.

Rather than scale the font based on the bounding box of the canvas, the
given font size is used as the basis for expanding the size of the box.
Larger fonts lead to larger displays. Smaller fonts lead to smaller
displays.

The API still needs to be worked out but the basic functionality works
by default.