Add `onClick` and `onMouseMove` event handlers to canvas terminal
This is a starting point for implementing UIs with mouse interactions.
The API isn’t fully worked out yet and there are probably some issues
relating to immediate mode vs retained mode style that will make it hard
to design a good API for this, but at least the basic functionality of
mapping mouse x/y to a given tile cell is there to build on top of.