Introduce `TextGrid` object to replace the `Terminal`
This is the first step in moving to a more structured API using
different types of grid as the core organising concept. The retro
terminal naming convention is neat, but kind of unnecessary here.

This changeset also deprecates the `Glyph` object in favour of `Cell`
which is a more meaningful name that better represents its role as the
main building block in the library.